# Swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Practice",
      "version": "1.0"
    },
    "paths": {
      "/Atheneum": {
        "get": {
          "tags": [
            "Atheneum"
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Atheneum"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Atheneum/{id}": {
        "get": {
          "tags": [
            "Atheneum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Atheneum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Atheneum"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Atheneum"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Book": {
        "get": {
          "tags": [
            "Book"
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Book"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/Book/{id}": {
        "get": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Book"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Atheneum": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32",
              "readOnly": true
            },
            "atheneumName": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "Book": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32",
              "readOnly": true
            },
            "bookName": {
              "type": "string",
              "nullable": true
            },
            "atheneum": {
              "type": "string",
              "nullable": true
            },
            "authors": {
              "type": "string",
              "nullable": true
            },
            "dateOfPublishing": {
              "type": "integer",
              "format": "int32"
            },
            "quantity": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }